# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BunkerWeb OpenAPI document, feel free to contact us if you have any questions or suggestions.",
        "title": "BunkerWeb OpenAPI",
        "contact": {}
    },
    "paths": {
...
        "/openapi/helper/get_user_state": {
            "post": {
                "description": "通过uid获取玩家状态信息, 例如主页点击数量, 需要提供 API key\n注意: 此 API 会尝试进行游戏登录 (Helper)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "玩家"
                ],
                "summary": "获取玩家状态信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "此处需要填写 API Key, 可以从用户菜单获取",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求时需要在 Body 携带以下查询参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/helper.GetUserStateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回",
                        "schema": {
                            "$ref": "#/definitions/helper.GetUserStateResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/helper/rental_server/search": {
            "post": {
                "description": "查询租赁服基础信息, 需要提供 API key\n注意: 此 API 会尝试进行游戏登录 (Helper)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租赁服 (查询类)"
                ],
                "summary": "查询租赁服信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "此处需要填写 API Key, 可以从用户菜单获取",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求时需要在 Body 携带以下查询参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental_server.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回",
                        "schema": {
                            "$ref": "#/definitions/rental_server.SearchResponse"
                        }
                    }
                }
            }
        },
        "/openapi/helper/search_user": {
            "post": {
                "description": "查询玩家, 可通过昵称来获取uid, 需要提供 API key\n注意: 此 API 会尝试进行游戏登录 (Helper)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "玩家"
                ],
                "summary": "查询玩家",
                "parameters": [
                    {
                        "type": "string",
                        "description": "此处需要填写 API Key, 可以从用户菜单获取",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求时需要在 Body 携带以下查询参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/helper.SearchUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回",
                        "schema": {
                            "$ref": "#/definitions/helper.SearchUserResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/owner/rental_server/create_backup": {
            "post": {
                "description": "为自己的服务器创建备份, 需要提供 API Key 且绑定游戏账号\n注意: 此 API 会尝试进行游戏登录 (Owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租赁服 (管理类)"
                ],
                "summary": "创建服务器备份",
                "parameters": [
                    {
                        "type": "string",
                        "description": "此处需要填写 API Key, 可以从用户菜单获取",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求时需要在 Body 携带以下查询参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental_server.CreateBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回",
                        "schema": {
                            "$ref": "#/definitions/giner.BasicResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/owner/rental_server/set_password": {
            "post": {
                "description": "设置自己服务器的密码, 需要提供 API Key 且绑定游戏账号\n注意: 此 API 会尝试进行游戏登录 (Owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租赁服 (管理类)"
                ],
                "summary": "设置服务器密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "此处需要填写 API Key, 可以从用户菜单获取",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求时需要在 Body 携带以下查询参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental_server.SetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回",
                        "schema": {
                            "$ref": "#/definitions/giner.BasicResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/owner/rental_server/turn_off": {
            "post": {
                "description": "关闭自己的服务器, 需要提供 API Key 且绑定游戏账号\n注意: 此 API 会尝试进行游戏登录 (Owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租赁服 (管理类)"
                ],
                "summary": "关闭服务器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "此处需要填写 API Key, 可以从用户菜单获取",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求时需要在 Body 携带以下查询参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental_server.TurnOffRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回",
                        "schema": {
                            "$ref": "#/definitions/giner.BasicResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/user/get_status": {
            "get": {
                "description": "用户信息查询, 需要提供 API Key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户中心"
                ],
                "summary": "用户信息查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "此处需要填写 API Key, 可以从用户菜单获取",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功时返回",
                        "schema": {
                            "$ref": "#/definitions/user.GetStatusResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bunker-web_routers_openapi_user.GetStatusResponseData": {
            "type": "object",
            "properties": {
                "create_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "expire_at": {
                    "description": "有效期至",
                    "type": "integer"
                },
                "game_id": {
                    "description": "绑定的游戏ID",
                    "type": "integer"
                },
                "is_admin": {
                    "description": "是否为系统管理员",
                    "type": "boolean"
                },
                "slots": {
                    "description": "Slots 列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bunker-web_routers_openapi_user.Slot"
                    }
                },
                "unlimited_until": {
                    "description": "无限制权限截止时间",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "bunker-web_routers_openapi_user.Slot": {
            "type": "object",
            "properties": {
                "expire_at": {
                    "description": "Slots 有效期至",
                    "type": "integer"
                },
                "game_id": {
                    "description": "Slots 绑定的游戏ID",
                    "type": "integer"
                },
                "note": {
                    "description": "Slots 备注",
                    "type": "string"
                }
            }
        },
        "g79.RentalServerInfo": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "chainInfo": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "has_pwd": {
                    "type": "string"
                },
                "icon_index": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "like_num": {
                    "type": "integer"
                },
                "mc_version": {
                    "type": "string"
                },
                "min_level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "offset": {
                    "description": "unknown type"
                },
                "owner_id": {
                    "type": "integer"
                },
                "player_count": {
                    "type": "integer"
                },
                "pvp": {
                    "type": "boolean"
                },
                "server_name": {
                    "type": "string"
                },
                "server_type": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "integer"
                },
                "world_id": {
                    "type": "string"
                }
            }
        },
        "giner.BasicResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "成功或失败原因",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        },
...
        "helper.GetUserStateRequest": {
            "type": "object",
            "properties": {
                "uid": {
                    "description": "目标玩家的启动器UID",
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "helper.GetUserStateResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "description": "成功或失败原因",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        },
        "helper.SearchUserRequest": {
            "type": "object",
            "properties": {
                "nickname": {
                    "description": "要查找的玩家昵称",
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "helper.SearchUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                },
                "message": {
                    "description": "成功或失败原因",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        },
...
        "rental_server.CreateBackupRequest": {
            "type": "object",
            "properties": {
                "backup_name": {
                    "description": "备份名称",
                    "type": "string",
                    "minLength": 1
                },
                "backup_slot": {
                    "description": "要使用的备份槽位, 范围为1-5",
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "server_code": {
                    "description": "服务器号",
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "rental_server.SearchRequest": {
            "type": "object",
            "properties": {
                "server_name": {
                    "description": "要查找的租赁服号",
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "rental_server.SearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/g79.RentalServerInfo"
                },
                "message": {
                    "description": "成功或失败原因",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        },
...
        "rental_server.SetPasswordRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "要设置的服务器密码, 内容为空则代表关闭密码",
                    "type": "string"
                },
                "server_code": {
                    "description": "服务器号",
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "rental_server.TurnOffRequest": {
            "type": "object",
            "properties": {
                "server_code": {
                    "description": "服务器号",
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "user.GetStatusResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/bunker-web_routers_openapi_user.GetStatusResponseData"
                },
                "message": {
                    "description": "成功或失败原因",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        },
...
    }
}